PR comments

+++ sdk/samples/iot/generate_certificate.ps1
@@ -1,14 +1,33 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # SPDX-License-Identifier: MIT
 
-openssl ecparam -out device_ec_key.pem -name prime256v1 -genkey
-openssl req -new -days 365 -nodes -x509 -key device_ec_key.pem -out device_ec_cert.pem -config x509_config.cfg -subj "/CN=paho-sample-device1"
-openssl x509 -noout -text -in device_ec_cert.pem
+# Check to make sure openssl is installed
+if (-Not (Get-Command "openssl" -ErrorAction SilentlyContinue)) {
+  throw "openssl is not availabe. You will need to install it or add it to the PATH to proceed"
+}
 
-if (Test-Path device_cert_store.pem)
-{
+# Remove previous cert stores
+if (Test-Path device_cert_store.pem) {
   Remove-Item device_cert_store.pem
 }
+
+#Wrap commands to catch errors
+function Invoke-IoTCommand {
+
+  param(
+    [string]$Command = $(throw "Must provide executable command string")
+  )
+
+  Invoke-Expression "$Command"
+
+  if ($LastExitCode -ne 0) {
+    throw "'$Command' invocation failed with exit code ($LastExitCode)"
+  }
+}
+
+Invoke-IoTCommand -Command 'openssl ecparam -out device_ec_key.pem -name prime256v1 -genkey'
+Invoke-IoTCommand -Command 'openssl req -new -days 365 -nodes -x509 -key device_ec_key.pem -out device_ec_cert.pem -config x509_config.cfg -subj "/CN=paho-sample-device1"'
+Invoke-IoTCommand -Command 'openssl -noout -text -in device_ec_cert.pem'
 
 Get-Content device_ec_cert.pem, device_ec_key.pem | Set-Content device_cert_store.pem
 
@@ -21,6 +40,7 @@ Write-Output "If using OpenSSL, it is recommended to use the OpenSSL Trusted CA 
 
 Write-Output "`nSAMPLE CERTIFICATE GENERATED:"
 Write-Output "Use the following command to set the environment variable for the samples:"
+
 if ($IsWindows) {
   Write-Output "`n`tset AZ_IOT_DEVICE_X509_CERT_PEM_FILE=$(Resolve-Path device_cert_store.pem)"
 }
@@ -34,6 +54,8 @@ Write-Output "Upload device_ec_cert.pem when enrolling your device with the Devi
 Write-Output "`nIOT HUB SAMPLES:"
 Write-Output "Use the following fingerprint when creating your device in IoT Hub."
 Write-Output "(The fingerprint has also been placed in fingerprint.txt for future reference.)"
+
 $fingerprint = openssl x509 -noout -fingerprint -in device_ec_cert.pem
-$fingerprint -replace ':','' | Tee-Object fingerprint.txt
+$fingerprint -replace ':', '' | Tee-Object fingerprint.txt
+
 Write-Output " "
